+++ src/main/kotlin/icons/SdkIcons.kt
@@ -1,0 +1,17 @@
+package icons
+
+import com.intellij.openapi.util.IconLoader
+
+/**
+ * https://plugins.jetbrains.com/docs/intellij/work-with-icons-and-images.html#how-to-organize-and-how-to-use-icons
+ * 必须在icons包下
+ * @author  qiubaisen
+ * @date  2021/3/3
+ */
+
+
+object SdkIcons {
+    @JvmField
+    val Sdk_default_icon = IconLoader.getIcon("/icons/sdk_16.svg")
+
+}