+++ CHANGELOG.md
@@ -3,5 +3,10 @@
 # intellij-plugin-demo Changelog
 
 ## [Unreleased]
+
+## [0.0.1]
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- Add Popup Dialog Action
+- Add Custom Action Group
+- Add Dynamic Action Group